Split AmbassadorConfig.py out.
